ofport and ofport_request operational data store persistence
Persists ofport and ofport_request fields in the operational data store.  The
yang model utilizes uint16 for these fields, while the library code utilizes
Long for consistency reasons.

+++ southbound/southbound-impl/src/main/java/org/opendaylight/ovsdb/southbound/ovsdb/transact/TerminationPointCreateCommand.java
@@ -56,11 +56,11 @@ public class TerminationPointCreateCommand implements TransactCommand {
             ovsInterface.setType(SouthboundMapper.createOvsdbInterfaceType(terminationPoint.getInterfaceType()));
             Integer ofPort = terminationPoint.getOfport();
             if (ofPort != null) {
-                ovsInterface.setOpenFlowPort(Sets.newHashSet(ofPort));
+                ovsInterface.setOpenFlowPort(Sets.newHashSet(ofPort.longValue()));
             }
             Integer ofPortRequest = terminationPoint.getOfportRequest();
             if (ofPortRequest != null) {
-                ovsInterface.setOpenFlowPortRequest(Sets.newHashSet(ofPortRequest));
+                ovsInterface.setOpenFlowPortRequest(Sets.newHashSet(ofPortRequest.longValue()));
             }
 
             //Configure optional input
